Add post new Genre and unit tests

+++ src/main/java/com/romeao/bookstore/api/v1/genre/GenreServiceImpl.java
@@ -1,13 +1,18 @@
 package com.romeao.bookstore.api.v1.genre;
 
 import com.romeao.bookstore.api.v1.util.Endpoints;
+import com.romeao.bookstore.api.v1.util.ErrorMessages;
 import com.romeao.bookstore.domain.Genre;
+import com.romeao.bookstore.errorhandling.ApiError;
+import com.romeao.bookstore.errorhandling.ApiException;
+import com.romeao.bookstore.errorhandling.ApiValidationError;
 import com.romeao.bookstore.repositories.GenreRepository;
 import com.romeao.bookstore.util.Link;
 import com.romeao.bookstore.util.LinkNames;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,6 +21,7 @@ import java.util.stream.Collectors;
 @Service
 public class GenreServiceImpl implements GenreService {
 
+    private static final Sort DEFAULT_SORT = Sort.by("name");
     private static final GenreMapper genreMapper = GenreMapper.INSTANCE;
     private final GenreRepository genreRepository;
 
@@ -32,24 +38,46 @@ public class GenreServiceImpl implements GenreService {
 
     @Override
     public List<GenreDto> findAll() {
-        return genreRepository.findAll(Sort.by("name")).stream()
+        return genreRepository.findAll(DEFAULT_SORT).stream()
                 .map(GenreServiceImpl::convertToDtoWithSelfLink)
                 .collect(Collectors.toList());
     }
 
     @Override
     public Page<GenreDto> findAll(int pageNum, int pageLimit) {
-        PageRequest page = PageRequest.of(pageNum, pageLimit, Sort.by("name"));
+        PageRequest page = PageRequest.of(pageNum, pageLimit, DEFAULT_SORT);
         return genreRepository.findAll(page).map(GenreServiceImpl::convertToDtoWithSelfLink);
-    }
-
-    @Override
-    public void deleteById(int genreId) {
-        genreRepository.deleteById(genreId);
     }
 
     @Override
     public GenreDto findById(int genreId) {
         return convertToDtoWithSelfLink(genreRepository.findById(genreId).orElse(null));
+    }
+
+    @Override
+    public GenreDto findByName(String genreName) {
+        return convertToDtoWithSelfLink(
+                genreRepository.findByNameIgnoreCase(genreName).orElse(null));
+    }
+
+    @Override
+    public void deleteById(int genreId) {
+        if (findById(genreId) == null) {
+            throw new ApiException(HttpStatus.NOT_FOUND, ErrorMessages.RESOURCE_NOT_FOUND);
+        }
+        genreRepository.deleteById(genreId);
+    }
+
+    @Override
+    public GenreDto save(GenreDto dto) {
+        if (findByName(dto.getName()) != null) {
+            ApiValidationError validationErr = new ApiValidationError("name",
+                    ErrorMessages.RESOURCE_EXISTS, dto.getName());
+            ApiError apiError = new ApiError(HttpStatus.UNPROCESSABLE_ENTITY,
+                    ErrorMessages.RESOURCE_EXISTS);
+            apiError.getSubErrors().add(validationErr);
+            throw new ApiException(apiError);
+        }
+        return convertToDtoWithSelfLink(genreRepository.save(genreMapper.toEntity(dto)));
     }
 }
